server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
python-multipart
Pillow
pyjwt[crypto]
passlib[bcrypt]
stripe
sqlalchemy